+++ modules/admin/views/ssh/index.php
@@ -1,9 +1,0 @@
-<?php
-/**
- * @author Semenov Alexander <semenov@skeeks.com>
- * @link http://skeeks.com/
- * @copyright 2010 SkeekS (СкикС)
- * @date 23.06.2015
- */
-?>
-<?= \skeeks\cms\widgets\ssh\SshConsoleWidget::widget(); ?>